Revert "Changed salary data type from int to double"
This reverts commit 902f314f8a3b854d44595569d9941b8ef2408b5c.

+++ src/main/java/org/gitsample/Employee.java
@@ -4,7 +4,7 @@ public class Employee {
 	//New comment added
 	private String name;
 	private int age;
-	private double salary;
+	private int salary;
 	private String city;
 
 	public String getName() {
